% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scipi.klnmf.R
\name{scipi.klnmf}
\alias{scipi.klnmf}
\title{SCI-PI for KL-NMF (SCale Invariant Power Iteration for KL-divergence NMF)}
\usage{
scipi.klnmf(
  V,
  k = 1,
  init = NULL,
  method = c("scipi", "em"),
  max.iter = 100,
  init.em.iter = 5,
  subprob.iter = 1,
  report.num = 1,
  verbose = FALSE
)
}
\arguments{
\item{X}{Describe X here. Sparse matrix? Dense?}

\item{y}{Describe y here.}
}
\value{
Describe return value here.
}
\description{

}
\details{
Add details here.
}
\examples{

}
